Revert "Feature/project management"
