+++ app/api/project/route.ts
@@ -2,48 +2,28 @@ import Project from "@/models/project";
 import { IProject } from "@/app/types/project";
 import { NextResponse } from "next/server";
 import { db } from "@/lib/db";
-import mongoose from "mongoose";
-import User from "@/models/user";
-import { formSchema } from "@/components/dashboard/forms/ProjectFormValidation";
 
-export async function GET() {
-  try {
+export async function GET () {
+  try
+  {
     await db();
     const projects = await Project.find();
     return NextResponse.json(projects, { status: 200 });
-  } catch (error) {
+  } catch (error)
+  {
     return NextResponse.json({ message: error }, { status: 500 });
   }
 }
 
-export async function POST(req: Request) {
-  try {
+export async function POST (req: Request) {
+  try
+  {
     await db();
     const body: IProject = await req.json();
-    const { name, startDate, endDate, tag, userId } = body;
-
-    if (!userId) {
-      return NextResponse.json(
-        { message: "User ID is required to create a project." },
-        { status: 400 }
-      );
-    }
-    console.log("userID:", body.userId);
-    const findUser = await User.findById(body.userId);
-
-    if (!findUser) {
-      return NextResponse.json({ message: "User not found." }, { status: 404 });
-    }
-
-    const newProject = await Project.create({
-      name,
-      startDate,
-      endDate,
-      tag,
-      userId,
-    });
-    return NextResponse.json(newProject, { status: 201 });
-  } catch (error) {
+    const project = await Project.create(body);
+    return NextResponse.json(project, { status: 201 });
+  } catch (error)
+  {
     return NextResponse.json({ message: error }, { status: 500 });
   }
 }
