--- app/dashboard/calendar/page.tsx
@@ -1,9 +1,0 @@
-import React from 'react'
-
-const page = () => {
-  return (
-    <div>calendar page</div>
-  )
-}
-
-export default page
