--- app/dashboard/members/page.tsx
@@ -1,7 +1,0 @@
-import React from "react";
-
-const page = () => {
-  return <div>members page</div>;
-};
-
-export default page;
