+++ app/page.tsx
@@ -1,13 +1,18 @@
 import RedirectButton from "@/components/RedirectBtn";
 import Dashboard from '@/components/dashboard/page';
 import { auth } from '@/lib/auth';
-import { redirect } from 'next/navigation';
 export default async function Home () {
   const session = await auth()
-  if (!session?.user) {
-    redirect('/')
-  }
   return (
-    <Dashboard />
+    <>
+      {session ? (
+        <Dashboard />
+      ) : (
+        <div>
+          <h1>You&apos;re not logged in</h1>
+          <RedirectButton />
+        </div>
+      )}
+    </>
   );
 }
