--- app/types/index.ts
@@ -1,12 +1,0 @@
-export enum FormFieldType {
-    INPUT = "input",
-    TEXTAREA = "textarea",
-    SELECT = "select",
-    RADIO = "radio",
-    CHECKBOX = "checkbox",
-    SWITCH = "switch",
-    Date_PICKER = "datePicker",
-    SKELETON = "skeleton",
-    PHONE_INPUT = "phoneInput",
-  }
-
