--- app/types/users.ts
@@ -1,9 +1,0 @@
-import { Types } from "mongoose";
-
-export interface IUser {
-    _id: Types.ObjectId;
-    name: string;
-    email: string;
-    image?: string;
-    emailVerified?: Date | null;
-}
