--- components/dashboard/forms/addProject.tsx
@@ -1,187 +1,0 @@
-"use client";
-import { useForm } from "react-hook-form";
-import { Form, FormControl } from "@/components/ui/form";
-
-import { zodResolver } from "@hookform/resolvers/zod";
-import { formSchema, ProjectFormSchema } from "./ProjectFormValidation";
-import { addProject } from "@/app/dashboard/actions";
-import { useRouter } from "next/navigation";
-import CustomFromField from "@/components/CustomFromField";
-import { FormFieldType } from "@/app/types";
-
-const AddProjectForm = () => {
-  const router = useRouter();
-  const form = useForm<ProjectFormSchema>({
-    resolver: zodResolver(formSchema),
-    defaultValues: {
-      name: "",
-      startDate: new Date(),
-      endDate: new Date(),
-      tag: "",
-    },
-  });
-
-  const onSubmit = async (values: ProjectFormSchema) => {
-    try {
-      await addProject(values);
-      console.log("Project added successfully!");
-      router.refresh();
-    } catch (error) {
-      console.log("Failed to add project.");
-    }
-  };
-
-  return (
-    <Form {...form}>
-      <form onSubmit={form.handleSubmit(onSubmit)} className="space-y-4">
-        <div>
-          <CustomFromField
-            control={form.control}
-            fieldType={FormFieldType.INPUT}
-            label="Project Name"
-            name="name"
-            placeholder="Project Name"
-            required
-          />
-          {form.formState.errors.name && (
-            <p className="text-red-500 text-sm mt-1">
-              {form.formState.errors.name.message}
-            </p>
-          )}
-        </div>
-        <div>
-          <CustomFromField
-            control={form.control}
-            fieldType={FormFieldType.Date_PICKER}
-            label="Start Date"
-            name="startDate"
-            required
-          />
-          {form.formState.errors.startDate && (
-            <p className="text-red-500 text-sm mt-1">
-              {form.formState.errors.startDate.message}
-            </p>
-          )}
-        </div>
-        <div>
-          <CustomFromField
-            control={form.control}
-            fieldType={FormFieldType.Date_PICKER}
-            label="End Date"
-            name="endDate"
-            required
-          />
-          {form.formState.errors.endDate && (
-            <p className="text-red-500 text-sm mt-1">
-              {form.formState.errors.endDate.message}
-            </p>
-          )}
-        </div>
-        <div>
-          <CustomFromField
-            control={form.control}
-            fieldType={FormFieldType.INPUT}
-            label="Tag (Optional)"
-            name="tag"
-            placeholder="Tag"
-          />
-          {form.formState.errors.tag && (
-            <p className="text-red-500 text-sm mt-1">
-              {form.formState.errors.tag.message}
-            </p>
-          )}
-        </div>
-        <button
-          type="submit"
-          className="w-full bg-blue-600 text-white py-2 px-4 rounded-md mt-4"
-        >
-          Add Project
-        </button>
-      </form>
-    </Form>
-    // <>
-    //   <form onSubmit={form.handleSubmit(onSubmit)} className="space-y-4">
-    //     <div>
-    //       <label htmlFor="name" className="block text-sm font-medium">
-    //         Project Name
-    //       </label>
-    //       <input
-    //         type="text"
-    //         id="name"
-    //         {...form.register("name")}
-    //         className="block w-full border rounded-md p-2 mt-1"
-    //       />
-    //       {form.formState.errors.name && (
-    //         <p className="text-red-500 text-sm mt-1">
-    //           {form.formState.errors.name.message}
-    //         </p>
-    //       )}
-    //     </div>
-
-    //     {/* Start Date Field */}
-    //     <div>
-    //       <label htmlFor="startDate" className="block text-sm font-medium">
-    //         Start Date
-    //       </label>
-    //       <input
-    //         type="date"
-    //         id="startDate"
-    //         {...form.register("startDate")}
-    //         className="block w-full border rounded-md p-2 mt-1"
-    //       />
-    //       {form.formState.errors.startDate && (
-    //         <p className="text-red-500 text-sm mt-1">
-    //           {form.formState.errors.startDate.message}
-    //         </p>
-    //       )}
-    //     </div>
-
-    //     {/* End Date Field */}
-    //     <div>
-    //       <label htmlFor="endDate" className="block text-sm font-medium">
-    //         End Date
-    //       </label>
-    //       <input
-    //         type="date"
-    //         id="endDate"
-    //         {...form.register("endDate")}
-    //         className="block w-full border rounded-md p-2 mt-1"
-    //       />
-    //       {form.formState.errors.endDate && (
-    //         <p className="text-red-500 text-sm mt-1">
-    //           {form.formState.errors.endDate.message}
-    //         </p>
-    //       )}
-    //     </div>
-
-    //     {/* Tag Field */}
-    //     <div>
-    //       <label htmlFor="tag" className="block text-sm font-medium">
-    //         Tag (Optional)
-    //       </label>
-    //       <input
-    //         type="text"
-    //         id="tag"
-    //         {...form.register("tag")}
-    //         className="block w-full border rounded-md p-2 mt-1"
-    //       />
-    //       {form.formState.errors.tag && (
-    //         <p className="text-red-500 text-sm mt-1">
-    //           {form.formState.errors.tag.message}
-    //         </p>
-    //       )}
-    //     </div>
-
-    //     {/* Submit Button */}
-    //     <button
-    //       type="submit"
-    //       className="w-full bg-blue-600 text-white py-2 px-4 rounded-md mt-4"
-    //     >
-    //       Add Project
-    //     </button>
-    //   </form>
-    // </>
-  );
-};
-
-export default AddProjectForm;
