+++ components/dashboard/page.tsx
@@ -9,9 +9,23 @@ import Header from "../Header";
 
 export const dynamic = "force-dynamic";
 
-export default function Dashboard () {
+export default function Dashboard() {
   const { data: session, status } = useSession();
   const router = useRouter();
+
+  useEffect(() => {
+    if (status === "unauthenticated") {
+      router.replace("/sign-in");
+    }
+  }, [status, router]);
+
+  if (status === "loading") {
+    return <p>Loading...</p>;
+  }
+
+  if (!session) {
+    return null; // Avoid rendering anything until redirection completes
+  }
 
   return (
     <div className="flex flex-col justify-start ">
