+++ lib/auth.ts
@@ -17,23 +17,6 @@ export const { handlers, auth, signIn, signOut } = NextAuth({
     allowDangerousEmailAccountLinking: true,
   })],
   pages: {
-    signIn: "/",
+    signIn: "/signin",
   },
-  callbacks: {
-    authorized: async ({ auth }) => {
-      // Logged in users are authenticated, otherwise redirect to login page
-      return !!auth
-    },
-    async jwt ({ token, user }) {
-      if (user) token.id = user.id;
-      return token;
-    },
-    async session ({ session, token }) {
-      if (token) {
-        session.user.id = token.id as string; // Map token `id` to session
-        session.user.email = token.email as string; // Map token `email` to session
-      }
-      return session;
-    },
-  }
 })
