+++ types/next-auth.d.ts
@@ -5,10 +5,6 @@ declare module "next-auth" {
   /**
    * Returned by `useSession`, `getSession` and received as a prop on the `SessionProvider` React Context
    */
-  interface User {
-    id: string,
-    email: string
-  }
   interface Session {
     user: {
       /** The user's postal address. */
